initial calibration works here

+++ xdrip/Transmitter/Generic/RawGlucoseData.swift
@@ -3,9 +3,9 @@ import Foundation
 /// raw glucose as received from transmitter
 struct RawGlucoseData {
     var timeStamp:Date
-    var glucoseLevelRaw:Int
+    var glucoseLevelRaw:Double
     
-    init(timeStamp:Date, glucoseLevelRaw:Int) {
+    init(timeStamp:Date, glucoseLevelRaw:Double) {
         self.timeStamp = timeStamp
         self.glucoseLevelRaw = glucoseLevelRaw
     }}
